Moved const into its own folder
