+++ src/components/forms/form-inputs/validate-mobile.js/validateCodeMobile.js
@@ -1,0 +1,7 @@
+const validatePhoneCode = ({ code }) => {
+  const errors = {};
+  if (String(code).length !== 6) errors.code = "Must be 6 characters";
+  return errors;
+};
+
+export default validatePhoneCode;
