+++ src/components/forms/form-inputs/validate-mobile.js/validateMobile.js
@@ -1,0 +1,10 @@
+import isMobilePhone from "validator/lib/isMobilePhone";
+
+const validatePhoneNumber = ({ mobile }) => {
+  const errors = {};
+  if (!isMobilePhone(mobile, "en-AU"))
+    errors.mobile = "Please enter a valid mobile number";
+  return errors;
+};
+
+export default validatePhoneNumber;
